Reorganise content of variables.py into several new modules (varcollection, variable_helpers, varnameinfo, variable)
