+++ pyaerocom/variable.py
@@ -1,10 +1,6 @@
 from ast import literal_eval
-from cf_units import Unit
 from configparser import ConfigParser
-import fnmatch
 import numpy as np
-import os
-import re
 
 from pyaerocom import logger, print_log, var_groups
 from pyaerocom.mathutils import make_binlist
@@ -13,255 +9,9 @@ from pyaerocom.exceptions import VariableDefinitionError
 from pyaerocom._lowlevel_helpers import list_to_shortstr, dict_to_str
 
 #: helper vor checking if variable name contains str 3d or 3D
-is_3d = lambda var_name: True if '3d' in var_name.lower() else False
-
-class VarNameInfo(object):
-    """This class can be used to retrieve information from variable names"""
-
-    #: valid number range for retrieval of wavelengths from variable name
-    _VALID_WVL_RANGE = [0.1, 10000] #nm
-
-    #: valid variable families for wavelength retrievals
-    _VALID_WVL_IDS = ['od', 'abs', 'ec', 'sc', 'ac', 'bsc', 'ssa']
-
-    PATTERNS = {'od' : r'od\d+aer'}
-    DEFAULT_VERT_CODE_PATTERNS = {
-        'abs*'  :   'Column',
-        'od*'   :   'Column',
-        'ang*'  :   'Column',
-        'load*' :   'Column',
-        'wet*'  :   'Surface',
-        'dry*'  :   'Surface',
-        'emi*'  :   'Surface'}
-
-    def __init__(self, var_name):
-        self.var_name = var_name
-        self._nums = []
-        try:
-            self._nums = self._numbers_in_string(var_name)
-        except Exception:
-            pass
-
-    def get_default_vert_code(self):
-        """Get default vertical code for variable name"""
-        for pattern, code in self.DEFAULT_VERT_CODE_PATTERNS.items():
-            if fnmatch.fnmatch(self.var_name, pattern):
-                return code
-        raise ValueError('No default vertical code could be found for {}'
-                         .format(self.var_name))
-
-    @staticmethod
-    def _numbers_in_string(input_str):
-        """Get list of all numbers in input str
-
-        Parameters
-        ----------
-        input_str : str
-            string to be checked
-
-        Returns
-        -------
-        list
-            list of numbers that were found in input string
-        """
-        return [int(x) for x in re.findall(r'\d+', input_str)]
-
-    @property
-    def contains_numbers(self):
-        """Boolean specifying whether this variable name contains numbers"""
-        if len(self._nums) > 0:
-            return True
-        return False
-
-    @property
-    def is_wavelength_dependent(self):
-        """Boolean specifying whether this variable name is wavelength dependent"""
-        for item in self._VALID_WVL_IDS:
-            if self.var_name.startswith(item):
-                return True
-        return False
-
-    @property
-    def contains_wavelength_nm(self):
-        """Boolean specifying whether this variable contains a certain wavelength"""
-        if not self.contains_numbers:
-            return False
-        low, high = self._VALID_WVL_RANGE
-        if self._nums and low <= self._nums[0] <= high:
-            return True
-        return False
-
-    @property
-    def wavelength_nm(self):
-        """Wavelength in nm (if appliable)"""
-        if not self.is_wavelength_dependent:
-            raise VariableDefinitionError('Variable {} is not wavelength '
-                                          'dependent (does not start with '
-                                          'either of {})'.format(self.var_name,
-                                                     self._VALID_WVL_IDS))
-
-        elif not self.contains_wavelength_nm:
-            raise VariableDefinitionError('Wavelength could not be extracted '
-                                          'from variable name')
-        return self._nums[0]
-
-    def in_wavelength_range(self, low, high):
-        """Boolean specifying whether variable is within wavelength range
-
-        Parameters
-        ----------
-        low : float
-            lower end of wavelength range to be tested
-        high : float
-            upper end of wavelength range to be tested
-
-        Returns
-        -------
-        bool
-            True, if this variable is wavelength dependent and if the
-            wavelength that is inferred from the filename is within the
-            specified input range
-        """
-        return low <= self.wavelength <= high
-
-    def translate_to_wavelength(self, to_wavelength):
-        """Create new variable name at a different wavelength
-
-        Parameters
-        ----------
-        to_wavelength : float
-            new wavelength in nm
-
-        Returns
-        -------
-        VarNameInfo
-            new variable name
-        """
-        if not self.contains_wavelength_nm:
-            raise ValueError('Variable {} is not wavelength dependent'.format(self.var_name))
-        name = self.var_name.replace(str(self.wavelength_nm),
-                                     str(to_wavelength))
-        return VarNameInfo(name)
-
-    def __str__(self):
-        s = ('\nVariable {}\n'
-             'is_wavelength_dependent: {}\n'
-             'is_optical_density: {}'.format(self.var_name,
-                                                self.is_wavelength_dependent,
-                                                self.is_optical_density))
-        if self.is_wavelength_dependent:
-            s += '\nwavelength_nm: {}'.format(self.wavelength_nm)
-        return s
-
-def parse_variables_ini(fpath=None):
-    """Returns instance of ConfigParser to access information"""
-    from pyaerocom import __dir__
-    if fpath is None:
-        fpath = os.path.join(__dir__, "data", "variables.ini")
-
-    if not os.path.exists(fpath):
-        raise FileNotFoundError("FATAL: variables.ini file could not be found "
-                                "at {}".format(fpath))
-    parser = ConfigParser()
-    parser.read(fpath)
-    return parser
-
-def parse_aliases_ini():
-    """Returns instance of ConfigParser to access information"""
-    from pyaerocom import __dir__
-    fpath = os.path.join(__dir__, "data", "aliases.ini")
-    if not os.path.exists(fpath):
-        raise FileNotFoundError("FATAL: aliases.ini file could not be found "
-                                "at {}".format(fpath))
-    parser = ConfigParser()
-    parser.read(fpath)
-    return parser
-
-
-def get_emep_variables(parser=None):
-    """Read variable definitions from emep_variables.ini file
-
-    Returns
-    -------
-    dict
-        keys are AEROCOM standard names of variable, values are EMEP variables
-    """
-    if parser is None:
-        parser = parse_emep_variables_ini()
-    variables = {}
-    items = parser['emep_variables']
-    for var_name in items:
-        _variables = [x.strip() for x in items[var_name].strip().split(',')]
-        for variable in _variables:
-            variables[var_name] = variable
-    return variables
-
-def parse_emep_variables_ini(fpath=None):
-    """Returns instance of ConfigParser to access information"""
-    from pyaerocom import __dir__
-    if fpath is None:
-        fpath = os.path.join(__dir__, "data", "emep_variables.ini")
-    if not os.path.exists(fpath):
-        raise FileNotFoundError("FATAL: emep_variables.ini file could not be found "
-                        "at {}".format(fpath))
-    parser = ConfigParser()
-    # added 12.7.21 by jgliss for EMEP trends processing. See here:
-    # https://stackoverflow.com/questions/1611799/preserve-case-in-configparser
-    parser.optionxform=str
-    parser.read(fpath)
-    return parser
-
-
-def _read_alias_ini(parser=None):
-    """Read all alias definitions from aliases.ini file and return as dict
-
-    Returns
-    -------
-    dict
-        keys are AEROCOM standard names of variable, values are corresponding
-        aliases
-    """
-    if parser is None:
-        parser = parse_aliases_ini()
-    aliases = {}
-    items = parser['aliases']
-    for var_name in items:
-        _aliases = [x.strip() for x in items[var_name].strip().split(',')]
-        for alias in _aliases:
-            aliases[alias] = var_name
-    for var_fam, alias_fam in parser['alias_families'].items():
-        if ',' in alias_fam:
-            raise Exception('Found invalid definition of alias family {}: {}. '
-                            'Only one family can be mapped to a variable name'
-                            .format(var_fam, alias_fam))
-    return aliases
-
-def get_aliases(var_name, parser=None):
-    """Get aliases for a certain variable"""
-    if parser is None:
-        from pyaerocom import __dir__
-        file = os.path.join(__dir__, "data", "aliases.ini")
-        parser = ConfigParser()
-        parser.read(file)
-
-    info = parser['aliases']
-    aliases = []
-    if var_name in info:
-        aliases.extend([a.strip() for a in info[var_name].split(',')])
-    for var_fam, alias_fam in parser['alias_families'].items():
-        if var_name.startswith(var_fam):
-            alias = var_name.replace(var_fam, alias_fam)
-            aliases.append(alias)
-    return aliases
-
-def _check_alias_family(var_name, parser):
-    for var_fam, alias_fam in parser['alias_families'].items():
-        if var_name.startswith(alias_fam):
-            var_name_aerocom = var_name.replace(alias_fam, var_fam)
-            return var_name_aerocom
-    raise VariableDefinitionError('Input variable could not be identified as '
-                                  'belonging to either of the available alias '
-                                  'variable families')
+from pyaerocom.variable_helpers import parse_variables_ini, parse_aliases_ini, \
+    _read_alias_ini, get_aliases, _check_alias_family
+from pyaerocom.varnameinfo import VarNameInfo
 
 class Variable(object):
     """Interface that specifies default settings for a variable
@@ -384,10 +134,20 @@ class Variable(object):
     def _check_input_var_name(var_name):
         if '3d' in var_name:
             var_name = var_name.replace('3d','')
+        elif '3D' in var_name:
+            var_name = var_name.replace('3D', '')
+        elif '_' in var_name:
+            raise ValueError(f'invalid variable name {var_name}. Must '
+                             f'not contain underscore')
         return var_name
 
-    def __init__(self, var_name="od550aer", init=True, cfg=None, **kwargs):
-        # save orig. input for whatever reasons
+    def __init__(self, var_name=None, init=True, cfg=None, **kwargs):
+        if var_name is None:
+            var_name = 'od550aer'
+        elif not isinstance(var_name, str):
+            raise ValueError(f'Invalid input for variable name, need str '
+                             f'type, got {type(var_name)}')
+        # save orig. input for whatever reason
         self._var_name_input = var_name
 
         self.var_name = self._check_input_var_name(var_name)
@@ -725,7 +485,9 @@ class Variable(object):
         """
         if cfg is None:
             cfg = self.read_config()
-
+        elif not isinstance(cfg, ConfigParser):
+            raise ValueError(f'invalid input for cfg, need config parser got '
+                             f'{type(cfg)}')
         if not var_name in cfg:
             try:
                 var_name = self._check_aliases(var_name)
@@ -811,228 +573,3 @@ class Variable(object):
 
         s += plot_s
         return s
-
-class VarCollection(object):
-    """Variable access class based on variables config file"""
-
-    def __init__(self, var_ini):
-        self._all_vars = None
-        self._var_ini = None
-
-        self.var_ini = var_ini
-
-        self._vars_added = {}
-
-        self._cfg_parser = parse_variables_ini(var_ini)
-        self._alias_parser = parse_aliases_ini()
-        self._idx = -1
-
-        logger.info("Importing variable aliases info")
-
-    @property
-    def all_vars(self):
-        """List of all variables
-
-        Note
-        ----
-        Does not include variable names that may be inferred via
-        alias families as defined in section [alias_families] in
-        aliases.ini.
-        """
-        if self._all_vars is None:
-            all_vars = [k for k in self._cfg_parser.keys()]
-            all_vars.extend(self._vars_added.keys())
-            #all_vars.extend(list(_read_alias_ini()))
-            self._all_vars = all_vars
-        return self._all_vars
-
-    @property
-    def var_ini(self):
-        """Config file specifying variable information"""
-        return self._var_ini
-
-    @var_ini.setter
-    def var_ini(self, var_ini):
-
-        if not os.path.exists(var_ini):
-            raise IOError("File {} does not exist".format(var_ini))
-        self._var_ini = var_ini
-
-    def find(self, search_pattern):
-        """Find all variables that match input search pattern
-
-        Note
-        ----
-        Searches for matches in variable names (:attr:`Variable.var_name`) and
-        standard name (:attr:`Variable.standard_name`).
-
-        Parameters
-        ----------
-        search_pattern : str
-            variable search pattern
-
-        Returns
-        -------
-        list
-            AeroCom variable names that match the search pattern
-        """
-        matches = []
-        for var in self:
-            if fnmatch.fnmatch(var.var_name, search_pattern):
-                matches.append(var.var_name)
-            elif (isinstance(var.standard_name, str) and
-                  fnmatch.fnmatch(var.standard_name, search_pattern)):
-                matches.append(var.var_name)
-        return matches
-
-    def get_coord_var_and_standard_names(self):
-        """Get dictionary with coord and standard names"""
-        d = {}
-        for k in self.all_vars:
-            d[k] = self[k]['standard_name']
-        return d
-
-    def _read_ini(self):
-        parser = ConfigParser()
-        parser.read(self.var_ini)
-        return parser
-
-    def __dir__(self):
-        """Activates auto tab-completion for all variables"""
-        return self.all_vars
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        self._idx += 1
-        if self._idx == len(self.all_vars):
-            self._idx = -1
-            raise StopIteration
-        var_name = self.all_vars[self._idx]
-        return self[var_name]
-
-    def __contains__(self, var_name):
-        if var_name in self.all_vars:
-            return True
-        return False
-
-    def __getattr__(self, attr):
-        if attr in self.__dict__:
-            return self.__dict__[attr]
-        return self[attr]
-
-    def add_var(self, var):
-        """Add a new variable to this collection
-
-        Minimum requirement for new variables are attributes var_name and
-        units.
-
-        Parameters
-        ----------
-        var : Variable
-            new variable definition
-
-        Raises
-        ------
-        VariableDefinitionError
-            if a variable is already defined under that name
-
-        Returns
-        -------
-        None
-        """
-        if not isinstance(var.var_name, str):
-            raise ValueError('Attr. var_name needs to be assigned to input '
-                             'variable')
-        if var.var_name in self.all_vars:
-            raise VariableDefinitionError(f'variable with name {var.var_name} '
-                                          f'is already defined')
-        if not isinstance(var, Variable):
-            raise ValueError('Can only add instances of Variable class...')
-        if not isinstance(var.units, str):
-            if not isinstance(var.units, Unit):
-                raise ValueError('Please assign a unit to the new input '
-                                 'variable')
-            var.units = str(var.units)
-        self._all_vars.append(var.var_name)
-        self._vars_added[var.var_name] = var
-
-    def get_var(self, var_name):
-        """
-        Get variable based on variable name
-
-        Parameters
-        ----------
-        var_name : str
-            name of variable
-
-        Raises
-        ------
-        VariableDefinitionError
-            if no variable under input var_name is registered.
-
-        Returns
-        -------
-        Variable
-            Variable instance
-
-        """
-        if var_name in self._vars_added:
-            return self._vars_added[var_name]
-        var = Variable(var_name, cfg=self._cfg_parser)
-        if not var.var_name_aerocom in self:
-            raise VariableDefinitionError(
-                'Error (VarCollection): input variable '
-                '{} is not supported'.format(var_name))
-        return var
-
-    def __getitem__(self, var_name):
-        return self.get_var(var_name)
-
-    def __repr__(self):
-        head = "Pyaerocom {}".format(type(self).__name__)
-        s = '\n{}\n{}\n{}'.format(len(head)*"-", head, len(head)*"-")
-        for v in self.all_vars:
-            s += '\n{}'.format(v)
-        return s
-
-    def __str__(self):
-        return self.var_name
-
-def get_variable(var_name):
-    """
-    Get a certain variable
-
-    Parameters
-    ----------
-    var_name : str
-        variable name
-
-    Returns
-    -------
-    Variable
-    """
-    from pyaerocom import const
-    return const.VARS[var_name]
-
-def all_vars_to_dataframe():
-    """Make an overview table for all variables"""
-    import pandas as pd
-    head = ['Name', 'Standard name', 'unit', 'Wavelength [nm]', 'Dimensions',
-            'Comments and purpose']
-    res = []
-    for varname in all_var_names():
-        var = Variable(varname)
-        res.append([varname, var.standard_name, var.unit, var.wavelength_nm,
-                    var.dimensions, var.comments_and_purpose])
-    df = pd.DataFrame(res, columns=head)
-    return df
-
-def all_var_names():
-    """Helper method that returns all currently defined variable names"""
-    return [k for k in Variable.read_config().keys()]
-
-if __name__=='__main__':
-    var = Variable('od550lt1aer')
-    print(var)
